Textpos fixup (#11)
* Fix broken builds when hyperref is not loaded

* Added full sans serif titlepage support

By adding the fonts-extra texlive collection to the container, we can use
libertine, which is a nice font for small capitals sans serif.

* Added proper fixup for wrong toc anchor

* Added options to tumcolor package

Plotcycle lists are defined in tumcolor already, so no need to have them in the
main file. Also add some more color names.

* Updated tumcolor documentation

* Added textpos fixup

Textpos is redefining shipout and thus clashes with some packages. Clearing the
textpos box seems to help avoiding clashes.

* Added some line termination comments

* Changed the hypersetup switch

* Fix hyperref setup

ifpackageloaded is allowed in the preambel only

* Fixed toc to toc switch